# _replit_import_20260506_153127/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ConnectSphere API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /venues:
    get:
      operationId: getVenues
      tags: [venues]
      summary: Get nearby venues from Google Places
      parameters:
        - name: lat
          in: query
          schema:
            type: number
          description: Latitude of search center
        - name: lng
          in: query
          schema:
            type: number
          description: Longitude of search center
        - name: radius
          in: query
          schema:
            type: integer
            default: 15000
          description: Search radius in meters (max 50000)
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category (All, Food & Drinks, Nightlife, Outdoors, Arts & Culture, Activities)
        - name: priceFilter
          in: query
          schema:
            type: string
          description: Filter by price tier (All, $, $$, $$$)
      responses:
        "200":
          description: List of nearby venues
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenuesResponse"
  /venues/saved:
    get:
      operationId: getSavedVenues
      tags: [venues]
      summary: Get the authenticated user's saved venue place IDs
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Saved venue IDs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedVenuesResponse"
  /venues/{placeId}/save:
    post:
      operationId: saveVenue
      tags: [venues]
      summary: Save a venue to the user's favourites
      security:
        - clerkAuth: []
      parameters:
        - name: placeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveVenueBody"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveVenueResponse"
    delete:
      operationId: unsaveVenue
      tags: [venues]
      summary: Remove a venue from the user's favourites
      security:
        - clerkAuth: []
      parameters:
        - name: placeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unsaved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveVenueResponse"
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    Venue:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
        priceLevel:
          type: integer
        priceTier:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        address:
          type: string
        rating:
          type: number
        photoUrl:
          type: string
        isOpen:
          type: boolean
        distance:
          type: number
      required:
        - id
        - name
        - category
        - priceLevel
        - priceTier
        - latitude
        - longitude
        - address
        - photoUrl
    VenuesResponse:
      type: object
      properties:
        venues:
          type: array
          items:
            $ref: "#/components/schemas/Venue"
        configured:
          type: boolean
      required:
        - venues
        - configured
    SaveVenueBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        address:
          type: string
        photoUrl:
          type: string
        priceTier:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        rating:
          type: number
      required:
        - name
        - category
        - address
        - photoUrl
        - priceTier
        - latitude
        - longitude
    SaveVenueResponse:
      type: object
      properties:
        saved:
          type: boolean
      required:
        - saved
    SavedVenuesResponse:
      type: object
      properties:
        placeIds:
          type: array
          items:
            type: string
      required:
        - placeIds